Module Chrome_trace

Output trace data to a file in Chrome's trace_event format. This format is compatible with chrome trace viewer chrome://tracing.

Trace viewer is a part of the catapult project (https://github.com/catapult-project/catapult/blob/master/tracing/README.md).

The trace format is documented at: https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview

alert unstable The API of this library is not stable and may change without notice.

module Json : sig ... end

module Id : sig ... end

module Stack_frame : sig ... end

module Event : sig ... end

module Output_object : sig ... end

The object format provided in whole